\name{ptb.frequency.domain}
\alias{ptb.frequency.domain}
\alias{ptb.cepstral.domain}
\alias{ptb.wavelet.domain}
\title{
ECG Data Transform Domains
}
\description{
These functions will transform ECG data into an additional domain (frequency, cepstral, wavelet).

}
\usage{
ptb.frequency.domain(vec, data.out=list(), block.size = 1024, sample.rate=1000)
ptb.cepstral.domain(vec, data.out=list(), block.size=1024, sample.rate=1000)
ptb.wavelet.domain(vec, data.out=list(), block.size=1024, sample.rate=1000)

}
\arguments{
  \item{vec}{
  A complex ECG vector, as generated by ptb.extract.lead.pair .
}
  \item{data.out}{
  A list object to fill with data in the new domain.
}
  \item{block.size}{
  The block size to use in analysis. For the frequency domain, this is the 'size' of the FFT.
}
  \item{sample.rate}{
  The rate at which the ECG data is sampled. This is generally ignored outside of the cepstral domain.
}
}
\details{
These functions are used to transform the time domain ECG data into other domains for subsequent analysis and feature generation.

The domain transform functions expect a complex vector; only ptb.feature.domain requires it.
}
\value{
The data.out object is filled with elements containing the transformed data.

For the frequency domain:
  \item{fft.a}{FFT for the dominant lead}
  \item{fft.b}{FFT for the subordinate lead}
  \item{auto.power.spectrum.a}{Auto Power Spectrum for the dominant lead}
  \item{auto.power.spectrum.b}{Auto Power Spectrum for the subordinate lead}
  \item{cross.power.spectrum}{Cross Power Spectrum}
  \item{cross.correlation}{Cross Correlation}
  
For the cepstral domain (using the tuneR package):
  \item{mfcc.a.cepstra}{Cepstral coefficients for the dominant lead}
  \item{mfcc.a.audio.spectrum}{Auditory spectrum for the dominant lead}
  \item{mfcc.a.power.spectrum}{Power spectrum for the dominant lead}
  \item{mfcc.b.cepstra}{Cepstral coefficients  for the subordinate lead}
  \item{mfcc.b.audio.spectrum}{Auditory spectrum for the subordinate lead}
  \item{mfcc.b.power.spectrum}{Power spectrum for the subordinate lead}
  
For the wavelet domain (using the wavelets package):
  \item{dwt.a.wave}{Wavelet coefficients for the dominant lead}
  \item{dwt.a.scale}{Scaling coefficients for the dominant lead}
  \item{dwt.b.wave}{Wavelet coefficients for the subordinate lead}
  \item{dwt.b.scale}{Wavelet coefficients for the subordinate lead}
}

\author{
mkfs (http://www.github.com/mkfs)
}
\note{
These transforms are experimental and are subject to future change.
}

\seealso{
  \code{\link{ptb.transform}}
}
\examples{
  data("patient.293")
  df.pair.239 <- ptb.extract.lead.pair(ptb.patient.293, 'v6', 'iii')
  ptb.lst <- apply(df.pair.239, 2, ptb.frequency.domain)
}
