\name{ptb.patient.293}
\alias{ptb.patient.293}
\alias{patient.293}
\docType{data}
\title{
PTB sample patient data
}
\description{
This dataset consists of a PTB list object for a single patient (293) from the Physionet PTB database. This patient has two test records.
}
\usage{data(ptb.patient.293)}
\format{
  The format is:
List of 2
 $ rec  :'data.frame':	2 obs. of  4 variables:
  ..$ ID     : num [1:2] 2930557 2930558
  ..$ Group  : chr [1:2] "s0557_re.dat" "s0558_re.dat"
  ..$ Patient: num [1:2] 293 293
  ..$ Age    : num [1:2] 75 75
 $ leads:List of 2
  ..$ 2930557:'data.frame':	120012 obs. of  15 variables:
  .. ..$ i  : num [1:120012] -2337 -3288 -4011 -3675 -3666 ...
  .. ..$ ii : num [1:120012] -3344 -3267 -3228 -3137 -3077 ...
  .. ..$ iii: num [1:120012] -1007 21 782 537 588 ...
  .. ..$ avr: num [1:120012] 2840 3277 3619 3405 3371 ...
  .. ..$ avl: num [1:120012] -666 -1656 -2398 -2107 -2128 ...
  .. ..$ avf: num [1:120012] -2176 -1623 -1222 -1299 -1244 ...
  .. ..$ v1 : num [1:120012] -72 147 263 60 -35 -217 -340 -517 -685 -958 ...
  .. ..$ v2 : num [1:120012] -1093 -688 -442 -498 -574 ...
  .. ..$ v3 : num [1:120012] -921 -470 -138 -117 -77 -114 -110 -180 -343 -488 ...
  .. ..$ v4 : num [1:120012] -1353 -843 -387 -309 -47 ...
  .. ..$ v5 : num [1:120012] -1858 -1320 -804 -724 -348 ...
  .. ..$ v6 : num [1:120012] -1292 -787 -318 -280 24 ...
  .. ..$ vx : num [1:120012] -2677 -2495 -2281 -2138 -1880 ...
  .. ..$ vy : num [1:120012] -1201 -1153 -1163 -1118 -1107 ...
  .. ..$ vz : num [1:120012] -1748 -1696 -1640 -1575 -1546 ...
  ..$ 2930558:'data.frame':	120012 obs. of  15 variables:
  .. ..$ i  : num [1:120012] 1217 1229 1255 1261 1234 ...
  .. ..$ ii : num [1:120012] 268 300 318 314 294 296 303 300 287 277 ...
  .. ..$ iii: num [1:120012] -947 -928 -936 -946 -938 -943 -950 -941 -924 -932 ...
  .. ..$ avr: num [1:120012] -742 -765 -787 -787 -764 -768 -779 -771 -749 -744 ...
  .. ..$ avl: num [1:120012] 1082 1079 1096 1104 1087 ...
  .. ..$ avf: num [1:120012] -338 -313 -308 -314 -321 -322 -322 -319 -317 -326 ...
  .. ..$ v1 : num [1:120012] -369 -365 -368 -368 -348 -345 -352 -346 -342 -339 ...
  .. ..$ v2 : num [1:120012] -225 -220 -219 -212 -206 -203 -217 -222 -217 -221 ...
  .. ..$ v3 : num [1:120012] 333 337 351 351 354 356 347 342 343 338 ...
  .. ..$ v4 : num [1:120012] -102 -93 -88 -97 -93 -86 -89 -99 -96 -95 ...
  .. ..$ v5 : num [1:120012] -39 -38 -32 -33 -38 -34 -39 -48 -42 -37 ...
  .. ..$ v6 : num [1:120012] -763 -759 -754 -753 -759 -765 -769 -778 -773 -768 ...
  .. ..$ vx : num [1:120012] -723 -727 -723 -727 -748 -746 -748 -758 -751 -749 ...
  .. ..$ vy : num [1:120012] -314 -297 -303 -310 -306 -301 -301 -293 -302 -322 ...
  .. ..$ vz : num [1:120012] -821 -825 -829 -822 -827 -827 -828 -835 -845 -849 ...
}
\source{
http://www.physionet.org/physiobank/database/ptbdb
http://www.physionet.org/physiobank/database/ptbdb/patient293/

}
\references{
Bousseljot R, Kreiseler D, Schnabel, A. Nutzung der EKG-Signaldatenbank CARDIODAT der PTB über das Internet. Biomedizinische Technik, Band 40, Ergänzungsband 1 (1995) S 317 

Goldberger AL, Amaral LAN, Glass L, Hausdorff JM, Ivanov PCh, Mark RG, Mietus JE, Moody GB, Peng C-K, Stanley HE. PhysioBank, PhysioToolkit, and PhysioNet: Components of a New Research Resource for Complex Physiologic Signals. Circulation 101(23):e215-e220 [Circulation Electronic Pages; http://circ.ahajournals.org/cgi/content/full/101/23/e215]; 2000 (June 13). 
}
\examples{
data(ptb.patient.293)
}
\keyword{datasets}
