\name{ptb.transform}
\alias{ptb.transform}
\alias{ptb.transform.vector}
\title{
Transform ECG into additional domains.
}
\description{
These functions will transform time-domain ECG data into additional domains (e.g. frequency, cepstral, wavelet).
}
\usage{
ptb.transform(ptb, lead.a = "v6", lead.b = "iii", block.size = 1024, sample.rate = 1000)
ptb.transform.vector(vec, block.size = 1024, sample.rate = 1000)

}
\arguments{
  \item{ptb}{
A PTB ist object
}
  \item{vec}{
A complex ECG vector, as generated by ptb.extract.lead.pair .
}
  \item{lead.a}{
The name of the dominant ('A') lead.
}
  \item{lead.b}{
The name of the subordinate ('B') lead.
}
  \item{block.size}{
The block size for use in analysis. For the frequency domain, this is the 'size' of the FFT.
}
  \item{sample.rate}{
The rate at which the ECG data is sampled. This is generally ignored outside of the cepstral domain.
}
}
\value{
The data.out object is filled with elements containing the transformed data.

See the documentation for \code{\link{ptb.frequency.domain}}, \code{\link{ptb.cepstral.domain}}, and \code{\link{ptb.wavelet.domain}}.
}
\author{
mkfs (http://www.github.com/mkfs)
}
\seealso{
\code{\link{ptb.frequency.domain}}
}
\examples{
  data("patient.293")
  ptb.data <- ptb.transform(ptb.patient.293)
  
  df.pair.239 <- ptb.extract.lead.pair(ptb.patient.293, 'v6', 'iii')
  ptb.lst <- apply(df.pair.239, 2, ptb.transform.vector)
}
